Move REST-related classes into the rest package
